Improve plugin and extension names
