+++ gradle-root-plugin/src/main/kotlin/io/spine/tools/gradle/root/plugin/RootExtension.kt
@@ -26,33 +26,30 @@
 
 package io.spine.tools.gradle.root.plugin
 
-import io.spine.tools.gradle.root.plugin.SpineSettingsExtension.Companion.NAME
-import org.gradle.api.Plugin
-import org.gradle.api.initialization.Settings
-import org.gradle.kotlin.dsl.create
+import org.gradle.api.plugins.ExtensionAware
 
 /**
- * Adds [spineSettings][SpineSettingsExtension] extension in the [Settings]
- * to which the plugin is applied.
+ * The root extension added by [RootPlugin].
  *
- * Before adding the extension, the plugin checks for the present of the extension.
- * So, applying the plugin more than once has no effect.
+ * This extension is used as a container for other extensions added by
+ * Gradle plugins of Spine libraries for configuration of compilation and other features.
+ *
+ * ### API Note
+ *
+ * Even though Gradle automatically makes a class created after the abstract extension
+ * class implement [ExtensionAware], we declare it explicitly to avoid the casts at
+ * the use sites in our code.
  */
-public class SpineSettingsPlugin : Plugin<Settings> {
-
-    override fun apply(settings: Settings) {
-        settings.run {
-            if (extensions.findByName(NAME) == null) {
-                extensions.create<SpineSettingsExtension>(NAME)
-            }
-        }
-    }
+public abstract class RootExtension : ExtensionAware {
 
     public companion object {
 
         /**
-         * The ID of the plugin.
+         * The name of the project extension.
+         *
+         * This is the name that starts the `spine { }` configuration block
+         * in a project to which [RootPlugin] is applied.
          */
-        public const val ID: String = "io.spine.settings"
+        public const val NAME: String = "spine"
     }
 }
